+++ ProjectEuler/Problems/1-10/Problem 005.fs
@@ -1,7 +1,7 @@
 ﻿module Problem_005
 
 let problem =
-    let productOfPrimes = [2.0..20.0] |> List.filter (int >> Common.isPrime) |> List.fold (fun acc e -> acc * e) 1.0
+    let productOfPrimes = [2.0..20.0] |> List.filter (int64 >> Common.isPrime) |> List.fold (fun acc e -> acc * e) 1.0
     let productOfAll = [2.0..20.0] |> List.fold (fun acc e -> acc * e) 1.0
     let seq = {productOfPrimes..productOfPrimes..productOfAll}
     seq
